backend/requirements.txt
flask==2.2.3
flask-cors==3.0.10
gunicorn==20.1.0
python-dotenv==1.0.0
numpy==1.24.2
pandas==1.5.3
torch==2.0.0
torchaudio==2.0.0
transformers==4.28.1
huggingface_hub==0.14.1
soundfile==0.12.1
google-generativeai==0.3.1
openai-whisper==20230314
plotly==5.14.1
Werkzeug==2.2.3